+++ src/main/java/sphabucks/products/vo/RequestSmallCategory.java
@@ -9,5 +9,5 @@ public class RequestSmallCategory {
     @Column(nullable = false)
     private String name;
     @Column(nullable = false)
-    private Integer bigCategoryId;
+    private Long bigCategoryId;
 }
