+++ src/main/java/sphabucks/users/vo/RequestUser.java
@@ -15,7 +15,7 @@ public class RequestUser {
     private String nickname;
     private String phoneNum;
     private String address;
-    private Integer star;
+    private Long star;
     private String carNum;
     private Date birth;
     private boolean sex;    // {true: M, false: F}
